Add Github Actions and fix wrangler.toml
