+++ .github/workflows/deploy-production.yaml
@@ -1,0 +1,26 @@
+name: Deploy Worker Production
+
+on:
+  push:
+    branches:
+      - main
+
+jobs:
+  production-deploy:
+    runs-on: ubuntu-latest
+    name: Deploy Production
+    steps:
+      - uses: actions/checkout@v2
+      - name: Yarn install with caching
+        uses: bahmutov/npm-install@v1.6.0
+
+      - name: Build Worker
+        run: yarn build
+      
+      - name: Publish Production
+        uses: cloudflare/wrangler-action@2.0.0
+        with:
+          apiToken: ${{ secrets.CF_API_TOKEN }}
+          accountId: ${{ secrets.CF_ACCOUNT_ID }}
+          workingDirectory: 'apps/worker'
+          command: publish --env=production
