+++ .github/workflows/deploy-staging.yaml
@@ -1,0 +1,27 @@
+name: Deploy Worker Staging
+
+on:
+  workflow_dispatch:
+  push:
+    branches-ignore:
+      - main
+
+jobs:
+  staging-deploy:
+    runs-on: ubuntu-latest
+    name: Deploy Staging
+    steps:
+      - uses: actions/checkout@v2
+      - name: Yarn install with caching
+        uses: bahmutov/npm-install@v1.6.0
+
+      - name: Build Worker
+        run: yarn build
+      
+      - name: Publish Staging
+        uses: cloudflare/wrangler-action@2.0.0
+        with:
+          apiToken: ${{ secrets.CF_API_TOKEN }}
+          accountId: ${{ secrets.CF_ACCOUNT_ID }}
+          workingDirectory: 'apps/worker'
+          command: publish --env=staging
