--- apps/front/.env.production
@@ -1,1 +1,0 @@
-VITE_WORKER_URL=https://api.knuclebones.io
