+++ apps/worker/src/types/env.d.ts
@@ -1,7 +1,6 @@
 export interface Env {
   ABLY_SERVER_SIDE_API_KEY: string
   ABLY_CLIENT_SIDE_API_KEY: string
-  SENTRY_DSN: string
   ABLY_JWT_STORE: KVNamespace
-  GAME_STATE_DURABLE: DurableObjectNamespace
+  GAME_STATE: DurableObjectNamespace
 }
